Created IRecordWriter & cosmetic changes
Created IRecordWriter and added a file for the compressed record writer
